Modification of stager generation :
Generate the stager in a custom directory

Before this commit stagers were generated in the SILENTTRINITY's root
directory. This commit adds an option allowing the user to choose the
location of the generated stager. This commit introduces a new directory
named "generated_stagers/" which is set as the new default directory for
stager generation.